Reduce the inner classes and enums, move many methods into the utility class
